requirements.txt
pyTelegramBotAPI
python-dotenv
Flask
gunicorn
psycopg2-binary


